Update InteractBuffer - Fix -Werror=sign-compare

+++ Includes/Interface/InteractBuffer.h
@@ -23,8 +23,8 @@ namespace Hearthstonepp
 
 		int GetCapacity() const;
 
-		size_t ReadBuffer(BYTE* data, int maxSize);
-		int WriteBuffer(BYTE* data, int size);
+		size_t ReadBuffer(BYTE* data, size_t maxSize);
+		size_t WriteBuffer(BYTE* data, size_t size);
 
 	private:
 		std::mutex m_mtx;
@@ -36,7 +36,7 @@ namespace Hearthstonepp
 		BYTE* m_buffer;
 		int m_head = 0;
 		int m_tail = 0;
-		int m_usage = 0;
+		size_t m_usage = 0;
 	};
 }
